Fix multiplayer sync issue

+++ code/src/multiplayer.h
@@ -52,7 +52,7 @@ void Multiplayer_Send_GSFlagBit(u8 index, u8 bit, u8 setOrUnset);
 void Multiplayer_Send_BigPoePoints(u32 pointDiff);
 void Multiplayer_Send_FishingFlag(u8 bit, u8 setOrUnset);
 void Multiplayer_Send_WorldMapBit(u8 bit, u8 setOrUnset);
-void Multiplayer_Send_ExtInfBit(u8 index, u8 bit, u8 setOrUnset);
+void Multiplayer_Send_ExtInfBit(u16 index, u8 bit, u8 setOrUnset);
 void Multiplayer_Send_TriforcePieces(u32 piecesDiff);
 void Multiplayer_Send_FullEntranceSync(u16 targetID);
 void Multiplayer_Send_DiscoveredScene(u32 index, u32 bit);
